# backend-node/openapi.yaml
openapi: 3.0.3
info:
  title: AI Services Platform API
  version: 2.0.0
  description: APIs for AI-powered services including Virtual Assistant, IDP, Computer Vision, Authentication, and more.
  contact:
    name: AI Services Platform Support
    email: support@aiservicesplatform.com
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.yourdomain.com
    description: Production server
...
paths:
...
  /api/chat/session:
    post:
      summary: Initialize a new chat session
      tags: [Chat Assistant]
      security:
        - cookieAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: string
                  description: Product ID (defaults to va-service)
                forceNew:
                  type: boolean
                  description: Force creation of new session even if active session exists
      responses:
        '200':
          description: Chat session initialized
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  customerId:
                    type: string
                  productId:
                    type: string
                  status:
                    type: string
                    enum: [initialized, resumed]
                  message:
                    type: string
                  messages:
                    type: array
                    description: Chat history (if resuming existing session)
                    items:
                      type: object
                  chatConfig:
                    type: object
                    properties:
                      greeting:
                        type: string
                      typingIndicator:
                        type: boolean
                      maxTurns:
                        type: integer
                      showIntent:
                        type: boolean
        '401':
          description: Unauthorized
        '403':
          description: Chat channel not enabled
        '502':
          description: Failed to start chat session
  /api/chat/message:
    post:
      summary: Send a message in an active chat session
      tags: [Chat Assistant]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
                message:
                  type: string
              required: [sessionId, message]
      responses:
        '200':
          description: Message processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  message:
                    type: string
                  intent:
                    type: string
                  extractedSlots:
                    type: object
                  requiresAction:
                    type: boolean
                  suggestedAction:
                    type: string
        '400':
          description: Missing required fields
        '502':
          description: Failed to process message
  /api/chat/end:
    post:
      summary: End a chat session
      tags: [Chat Assistant]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sessionId:
                  type: string
              required: [sessionId]
      responses:
        '200':
          description: Chat session ended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Missing sessionId
  /api/chat/history/{sessionId}:
    get:
      summary: Get conversation history for a session
      tags: [Chat Assistant]
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chat history
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        speaker:
                          type: string
                          enum: [user, assistant]
                        text:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
        '404':
          description: Session not found

  # Usage Endpoints
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie for authentication
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...